% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heating_advice.R
\name{heating_advice}
\alias{heating_advice}
\title{Heating_advice: should you increase or decrease your thermostat?}
\usage{
heating_advice(feeling, thermostat)
}
\arguments{
\item{feeling}{specify whether you are feeling hot or cold, string inputs "Hot" or "Cold"}

\item{thermostat}{specify the temperature that your thermostat is currently set to, numeric inputs}
}
\description{
This function will tell you how to adjust your temperature depending on what temperature your thermostat is set to, and if you are feeling hot or cold
}
\examples{
heating_advice(feeling = "cold", thermostat = 17)

}
