% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_advice.R
\name{window_advice}
\alias{window_advice}
\title{Window_advice: should you open or close windows?}
\usage{
window_advice(feeling, windows)
}
\arguments{
\item{feeling}{specify whether you are feeling hot or cold, string inputs "Hot" or "Cold"}

\item{windows}{specify how many windows you currently have open, numeric inputs}
}
\description{
This function will tell you how to adjust your temperature depending on how many windows you have open, and if you are feeling hot or cold
}
\examples{
window_advice(feeling = "hot", windows = 1)

}
